moved actions and reducer into own folders
