+++ client/src/Store/Redux/User/reducer/index.ts
@@ -1,0 +1,1 @@
+export * from './user.reducer';
