+++ client/src/Store/Redux/User/reducer/user.reducer.ts
@@ -5,7 +5,7 @@ import {
   REQUEST_READ_USER_INFO,
   SUCCESS_READ_USER_INFO,
   FAIL_READ_USER_INFO
-} from './user.types';
+} from '../user.types';
 
 export interface UserState extends User {
   loading: {
